requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
appdirs==1.4.4
    # via
    #   pyppeteer
    #   requests-cache
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via
    #   cattrs
    #   jsonschema
    #   referencing
    #   requests-cache
babel==2.11.0
    # via -r requirements.in
bcp-reader==0.1.1
    # via -r requirements.in
beautifulsoup4==4.11.2
    # via bs4
bleach==6.0.0
    # via django-sql-dashboard
boto3==1.33.6
    # via -r requirements.in
botocore==1.33.6
    # via
    #   boto3
    #   s3transfer
bs4==0.0.1
    # via requests-html
cattrs==22.2.0
    # via requests-cache
certifi==2023.7.22
    # via
    #   -r requirements.in
    #   elasticsearch
    #   pyppeteer
    #   requests
    #   sentry-sdk
charity-django @ git+https://github.com/kanedata/charity-django.git@v0.4.1
    # via -r requirements.in
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   click-default-group-wheel
    #   sqlite-utils
click-default-group-wheel==1.2.2
    # via sqlite-utils
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
confusable-homoglyphs==3.2.0
    # via django-registration
coverage==7.1.0
    # via -r requirements.in
cssselect==1.2.0
    # via pyquery
decorator==5.1.1
    # via validators
dj-database-url==1.2.0
    # via
    #   -r requirements.in
    #   charity-django
django==4.2.2
    # via
    #   -r requirements.in
    #   charity-django
    #   dj-database-url
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   django-markdownx
    #   django-ninja
    #   django-ninja-apikey
    #   django-prettyjson
    #   django-registration
    #   django-sql-dashboard
django-better-admin-arrayfield==1.4.2
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-debug-toolbar==3.8.1
    # via -r requirements.in
django-elasticsearch-dsl==7.2.2
    # via -r requirements.in
django-enforce-host==1.1.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-markdownx==4.0.0b1
    # via -r requirements.in
django-ninja==1.1.0
    # via
    #   -r requirements.in
    #   django-ninja-apikey
django-ninja-apikey==0.1.0
    # via -r requirements.in
django-prettyjson==0.4.1
    # via -r requirements.in
django-registration==3.3
    # via -r requirements.in
django-sql-dashboard @ git+https://github.com/drkane/django-sql-dashboard.git@dk/issue-141
    # via -r requirements.in
django-widget-tweaks==1.4.12
    # via -r requirements.in
elasticsearch==7.17.8
    # via elasticsearch-dsl
elasticsearch-dsl==7.4.0
    # via django-elasticsearch-dsl
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.1.3
    # via
    #   cattrs
    #   pytest
fake-useragent==1.1.1
    # via requests-html
faker==16.7.0
    # via charity-django
gunicorn==20.1.0
    # via -r requirements.in
humanize==4.5.0
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==6.0.0
    # via pyppeteer
inflect==7.0.0
    # via
    #   -r requirements.in
    #   charity-django
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.21.1
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
lml==0.1.0
    # via pyexcel-io
lxml==4.9.2
    # via
    #   -r requirements.in
    #   pyexcel-ezodf
    #   pyexcel-ods3
    #   pyquery
markdown==3.4.1
    # via
    #   django-markdownx
    #   django-sql-dashboard
markupsafe==2.1.2
    # via jinja2
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
openpyxl==3.1.0
    # via -r requirements.in
packaging==23.2
    # via pytest
parse==1.19.0
    # via requests-html
pillow==9.4.0
    # via django-markdownx
pluggy==1.4.0
    # via pytest
psycopg2==2.9.5
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via charity-django
pycountry==22.3.5
    # via -r requirements.in
pydantic==2.6.0
    # via
    #   -r requirements.in
    #   django-ninja
    #   inflect
pydantic-core==2.16.1
    # via pydantic
pyee==8.2.2
    # via pyppeteer
pyexcel-ezodf==0.3.4
    # via pyexcel-ods3
pyexcel-io==0.6.6
    # via pyexcel-ods3
pyexcel-ods3==0.6.1
    # via -r requirements.in
pyppeteer==1.0.2
    # via requests-html
pyquery==2.0.0
    # via requests-html
pytest==8.0.0
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   charity-django
    #   elasticsearch-dsl
    #   faker
    #   sqlite-utils
python-dotenv==0.21.1
    # via -r requirements.in
pytz==2022.7.1
    # via babel
pyyaml==6.0
    # via -r requirements.in
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements.in
    #   charity-django
    #   requests-cache
    #   requests-html
    #   requests-mock
    #   requests-oauthlib
    #   tweepy
requests-cache==0.9.8
    # via
    #   -r requirements.in
    #   charity-django
requests-html==0.10.0
    # via
    #   -r requirements.in
    #   charity-django
requests-mock==1.10.0
    # via -r requirements.in
requests-oauthlib==1.3.1
    # via tweepy
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
ruff==0.1.3
    # via -r requirements.in
s3transfer==0.8.2
    # via boto3
sentry-sdk==1.15.0
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   django-elasticsearch-dsl
    #   django-prettyjson
    #   elasticsearch-dsl
    #   python-dateutil
    #   requests-mock
    #   url-normalize
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlite-fts4==1.0.3
    # via sqlite-utils
sqlite-utils==3.30
    # via -r requirements.in
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
standardjson==0.3.1
    # via django-prettyjson
tabulate==0.9.0
    # via sqlite-utils
titlecase==2.4
    # via
    #   -r requirements.in
    #   charity-django
tomli==2.0.1
    # via pytest
tqdm==4.64.1
    # via
    #   -r requirements.in
    #   charity-django
    #   pyppeteer
tweepy==4.12.1
    # via -r requirements.in
typing-extensions==4.9.0
    # via
    #   inflect
    #   pydantic
    #   pydantic-core
tzdata==2022.7
    # via django
ua-parser==0.16.1
    # via -r requirements.in
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.14
    # via
    #   -r requirements.in
    #   botocore
    #   elasticsearch
    #   pyppeteer
    #   requests
    #   requests-cache
    #   sentry-sdk
validators==0.20.0
    # via -r requirements.in
w3lib==2.1.1
    # via requests-html
webencodings==0.5.1
    # via bleach
websockets==10.4
    # via pyppeteer
whitenoise==6.3.0
    # via -r requirements.in
xlsxwriter==3.1.9
    # via -r requirements.in
zipp==3.12.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
